Restructrured preprocessing scripts
